feat: add a complex type for the SQL handle
The previous approach was using a simple type to return the SQL handle.
The returned handle value was a mix between the result value of SQLite
operation and the SQL handle, where negative number indicates an error,
causing the issues in the recent devices running Android 11,
where the memory addresses of the handle were valid but they were also
negative numbers when converted to the `long long` type.

With this new complex type for the SQL Handle, we are now able to return
both the result of the SQLite operation and the SQL handle.

References https://outsystemsrd.atlassian.net/browse/RNMT-4515